requirements.txt
Flask>=3.0.0
python-dotenv>=1.0.0
torch>=2.0.0
torchvision>=0.15.0
open-clip-torch>=2.20.0
pillow>=10.0.0
numpy>=1.24.0
